[Refactor] RHF 요소 분리했습니다
