+++ src/api/products/client/usePutModifyProduct.ts
@@ -137,6 +137,7 @@ export const usePutModifyProduct = () => {
           context.previousProductList
         );
       }
+      alert("상품 수정에 실패했습니다. 다시 시도해주세요.");
     },
 
     onSettled: (data, error, variables) => {
